Created new subsystem for logging and debugging.
This commit adds a unified logging and debugging infrastructure. It is intended
to replace all other approaches and special command line flags. To use the
framework a modules should define the DEBUG_NAME macro.

#define DEBUG_NAME "mymodule"

The framework provides the macro DEBUG( expr ). expr is only executed if the module,
specified by DEBUG_NAME, is selected for debugging. This can be done via the
-XX:DebugName=<name> command line option. Note that is sufficient that <name>
is a suffix of DEBUG_NAME. So -XX:DebugName=jit will activate all modules
where DEBUG_NAME starts with "jit", eg "jit", "jit-compiler", "jit-inlining", etc.
This can be used to create debugging hierarchies.

The macro LOG( expr ) is like DEBUG() but it applies the expr to OStream object. It
also sets the prefix of the stream to DEBUG_NAME.

#define DEBUG_NAME "mymodule"
...
LOG("compiling " << method << " with flags: " << flags);

This will produce an output similar to (given that method and flags support OStream):

mymodule compiling java/lang/String with flags: [inlining]


* configure.ac: added --enable-logging
* m4/logging.m4: likewise
* src/toolbox/logging.cpp: add new logging framework
* src/toolbox/logging.hpp: likewise
* src/toolbox/Debug.cpp: added debugging framework
* src/toolbox/Debug.hpp: likewise
* src/toolbox/Makefile.am: likewise
* src/toolbox/OStream.cpp: minor adoptions
* src/toolbox/OStream.hpp: likewise
* src/vm/options.c: added -XX:DebugName=<name> command line option