logging: DEBUG* macros now enabled with --enable-logging configure flag
There was a discrepancy between the LOG* and DEBUG* macros. Until now
the LOG* stuff was enabled with --enable-logging (ENABLE_LOGGING define)
but DEBUG* was enabled with --enable-debug (not defined NDEBUG).

The DEBUG() macro is usually used if something can not be <<'ed into an
OStream. The following code illustrates the problem:

<code>
OStream  &print_complicated_stuff(OStream &OS) {
  return OS << "debug";
}

...

LOG("log" << nl);
DEBUG(print_complicated_stuff( dbg() ));
</code>


If both flags are enabled the output is:
<out>
log
debug
</out>

If debug is disabled but logging enabled:

<out>
log
</out>

And if logging disabled and debugging enabled:

<out>
debug
</out>


This is not the expected behaviour. Therefor all DEBUG*, LOG* are now enabled
by the --enable-logging configure flag.

Assertions, and other expensive integrity tests, are enabled by --enable-debug
(undef NDEBUG). The configure flag name is quite confusing in my eyes.
--enable-assertions would be more appropriate.


* src/toolbox/Debug.hpp: changed from !NDEBUG to ENABLE_LOGGING