doxygen: added footnote support
This commit adds some sort of footnote support to doxygen. The @footenote{x}
command is used to define a label. The @fn{x} is can be used to add references
to a footnote. It is intended as a doxygen aware replacement for the [1], [2],
etc. markers.
Example:

/**

  The algorithm continues to follow all possible paths @fn{1} until the
  recorded types do not change anymore.


  --- Footnotes

  @footnote{1} Actually the algorithm iterates through the basic block list until
  there are no more changes. Theoretically it would be wise to sort the
  basic blocks topologically beforehand, but the number of average/max
  iterations observed is so low, that this was not deemed necessary.

*/

Note: The label is not restricted to numbers.


* doc/doxygen/Doxyfile.in: added @fn and @footnote
* src/vm/jit/verify/typecheck.cpp: adopt @fn and @footnote