added future library support
This commit adds support C++ library features not present in C++98. As an
example support for shared_ptr has been added.

Autoconf tests for several implementations[1], in the case of shared_ptr for
std::tr1:shared_ptr and boost::shared_ptr. The best fit is made available via
cacao::shared_ptr.

Also note that the lack of a shared_ptr implementation does not necessarily
lead to a build error. As long as "toolbox/future.hpp" is not included
everything is fine.

[1]: another possibility is to enable C++11 and use std::shared_ptr but that
would change the behavior of several standard library features what in turn
may lead to issues. As long as the code is not C++11 ready this is not a
feasible option.

* configure.ac: check for future library support
* m4/future.m4: added
* src/toolbox/future.hpp: likewise