added support for gtest (google test xUnit testing framework)
Google test [1] is a C++ unit testing framework based on the xUnit architecture.
To enable it configure with `--with-gtest=<path>` where `<path>` points to a
gtest build. It is disabled by default.

To add a new unit test create a file `<name>.cpp` in the `tests/gtest` folder and add
`<name>` to the `TESTS` variable in `tests/gtest/Makefile.am`.

[1]: http://code.google.com/p/googletest/

* configure.ac: likewise
* tests/Makefile.am: likewise
* m4/gtest.m4: added
* tests/gtest/Makefile.am: likewise
* tests/gtest/libgtest_main.cpp: likewise
* tests/gtest/sample.cpp: likewise