Made it possible to compile cacao with C++11 and C11
There now are two new flags (--with-c-standard, --with-cpp-standard) with which you can set the C/C++ language standard to use (gcc/clang only).
Building with C++98 and C99 is still the default.

The whitespace changes in vm.cpp and properties.cpp are required by new syntax rules for string literals in C++11.

Note that including both fdlibm.h and cmath with C++11 leads to build errors,
because they have different prototypes for functions with the same name.
Since many STL headers depend on cmath they can't be included in a file that uses fdlibm.
Fortunately this can be avoided everywhere, but we might consider name-prefixing fdlibm functions.

Also note that m4/c-cxx-language-version.m4 has no name header on purpose.